Documenting Main Execution functions and fix of upscale_video
Part1 and 2 of the main execution cells don't finish once the cell execution has finished due to waiting on user selection. Documenting the waiting times.
upscale_video downloads the output file from GCS as it is removed locally when it was uploaded to GCS to be able to upscale it.