Add the framework for the backend code.
This adds a lot of the structure for the backend code, to allow the code to be migrated out of the colab.

The colab can then directly install the backend code from Github via pip.

This creates a simpler colab file, and provides more tooling for testing code, and extension to a UI in the future.

Change-Id: Id04ca6f0f8816bfce3e6567d98e03a0e522c2314